refactor: Move ICalGenerator logic to reusable ExportCustomizerICal component
