+++ pages/ICalGenerator.vue
@@ -1,221 +1,16 @@
 <template>
   <v-container>
     <WarnNoSetStreet/>
+    <h1 class="text-h3 text-center">{{i18n.t("iCalGen.pageTitle")}}</h1>
+    <p>{{ i18n.t("iCalGen.pageSubtitle") }}</p>
 
-  <h1 class="text-h3 text-center">{{i18n.t("iCalGen.pageTitle")}}</h1>
-
-  <p>{{ i18n.t("iCalGen.pageSubtitle") }}</p>
-
-  <v-card
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.pickupEventsTitle')"
-    :subtitle="i18n.t('iCalGen.pickupEventsSubtitle')"
-  >
-    <v-card-text>
-      <v-row>
-        <v-col cols="auto">
-          <EventTypeSelector v-model="formatOptions.type" :show-label="true"/>
-        </v-col>
-      </v-row>
-    </v-card-text>
-
-  </v-card>
-
-  <v-card
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.eventTimesTitle')"
-    :subtitle="i18n.t('iCalGen.eventTimesSubtitle')"
-  >
-    <v-card-text class="my-2">
-      <v-row>
-        <v-col cols="auto">
-          <v-input
-            :hide-details="true"
-            class="flex-0-1"
-          >
-            <TimeRangeInput 
-              v-model:start-time="formatOptions.startTime"
-              v-model:end-time="formatOptions.endTime"
-              :icon="mdiClockOutline"
-            />
-          </v-input>
-        </v-col>
-
-        <v-col cols="auto" align-self="center">
-          <v-divider vertical length="2em" opacity="50"/>
-        </v-col>
-
-        <v-col cols="auto">
-          <v-checkbox
-            v-model="formatOptions.allDay"
-            :hide-details="true"
-          >
-            <template #label>
-              <v-icon :icon="mdiHours24"/>
-              <span>{{ i18n.t('iCalGen.labelAllDay') }}</span>
-            </template>
-          </v-checkbox>
-        </v-col>
-
-      </v-row>
-    </v-card-text>
-  </v-card>
-
-  <v-card
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.eventAlarmTitle')"
-    :subtitle="i18n.t('iCalGen.eventAlarmSubtitle')"
-  >
-    <v-card-text>
-      <v-row>
-        <v-col cols="auto">
-          <v-input 
-            class="flex-0-1"
-            :hide-details="true"
-          >
-            <NumberInput
-              id="numberinput__alarm"
-              v-model:input-number="formatOptions.alarm"
-              :icon="mdiBellOutline"
-              :label="i18n.t('iCalGen.labelAlarmMinutesBefore')"
-              :min=0
-              :max=720
-              :step=5
-              :disabled="disabledAlarm"
-            />
-          </v-input>
-        </v-col>
-
-        <v-col cols="auto" align-self="center">
-          <v-divider vertical length="2em" opacity="50"/>
-        </v-col>
-
-        <v-col cols="auto">
-          <v-checkbox
-            v-model="disabledAlarm"
-            :hide-details="true"
-          >
-          <template #label>
-            <v-icon :icon="mdiBellOffOutline"/>
-            <span>{{i18n.t('iCalGen.labelAlarmDisable')}}</span>
-          </template>
-
-          </v-checkbox>
-        </v-col>
-      </v-row>
-    </v-card-text>
-  </v-card>
-
-  <v-card 
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.eventOffsetTitle')"
-    :subtitle="i18n.t('iCalGen.eventOffsetSubtitle')"
-  >
-
-    <v-card-text>
-      <v-row>
-        <v-col>
-          <v-input class="flex-0-1">
-            <NumberInput
-              :id="'numberinput__offset'"
-              v-model:input-number="formatOptions.offsetEvent"
-              :icon="mdiRewindOutline"
-              :label="i18n.t('iCalGen.labelOffsetDaysEarlier')"
-              :min=0
-              :max=7
-            />
-          </v-input>
-        </v-col>
-      </v-row>
-
-      <v-alert
-        color="info"
-        icon="$info"
-        title="Info"
-        :text="i18n.t('iCalGen.eventOffsetInfo')"
-      />
-    </v-card-text>
-  </v-card>
-
-  <v-card
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.eventTitleTitle')"
-    :subtitle="i18n.t('iCalGen.eventTitleSubtitle')"
-  >
-    <v-card-text>
-      <v-text-field
-        v-model="formatOptions.customSummary"
-        :label="i18n.t('iCalGen.labelEventTitle')"
-      />
-
-      <v-text-field
-        v-model="eventSummaryPreview"
-        :label="i18n.t('iCalGen.labelEventTitlePreview')"
-        :readonly="true"
-      />
-  
-      <v-alert
-        color="info"
-        icon="$info"
-        title="Info"
-      >
-        <p> {{ i18n.t('iCalGen.eventTitlePlaceholderInfo') }}</p>
-      </v-alert>
-
-    </v-card-text>
-  </v-card>
-
-  <v-card
-    class="my-4 pa-4"
-    :title="i18n.t('iCalGen.yourIcalLink')"
-  >
-    <v-card-text>
-      <v-row wrap="nowrap">
-        <v-col cols="auto">
-          <CopyToClipboardButton :text-to-copy="icalUrl.toString()"/>
-        </v-col>
-        <v-col>
-          <v-text-field
-            v-model="icalUrl"
-            class="h-25"
-            :editable="false"
-            :readonly="true"
-            :hide-details="true"
-          />
-        </v-col>
-
-      </v-row>
-
-    </v-card-text>
-  </v-card>
-
-</v-container>
-
+    <ExportCustomizerICal/>
+  </v-container>
 </template>
 
 <script setup lang="ts">
-  import TimeRangeInput from "~/components/TimeRangeInput.vue";
-  import CopyToClipboardButton from "~/components/CopyToClipboardButton.vue";
-  import EventTypeSelector from "~/components/EventTypeSelector.vue";
-  import { useCookieUserConfig } from "~/composables/useCookieUserConfig";
-  import type { ICalOptions } from "~/server/services/TransformDataService";
-  import type { AwsApiServiceEventTypeName } from "~/server/services/AwsApiService";
+  import ExportCustomizerICal from "~/components/ExportCustomizer/ExportCustomizerICal.vue";
 
-  import { 
-    mdiClockOutline, 
-    mdiHours24, 
-    mdiRewindOutline, 
-    mdiBellOutline,
-    mdiBellOffOutline
-  } from "@mdi/js";
-
-  type UiICalFormatOptions = Omit<ICalOptions, "startTime" | "endTime" | "translated"> & {
-    startTime?: string;
-    endTime?: string;
-    type?: AwsApiServiceEventTypeName[];
-  }
-
-  const { cookieStreet, cookieLanguage } = useCookieUserConfig();
   const { i18n, multiMergeLocaleMessage }  = useCustomI18n();
 
   multiMergeLocaleMessage("iCalGen", [
@@ -226,99 +21,5 @@
         en: "Use the options below to generate a custom iCal Calendar link for the pickup dates, that you can import into your calendar application"
       }
     ],
-    ["pickupEventsTitle", {de: "Abholungen", en: "Pickup Events"}],
-    [
-      "pickupEventsSubtitle", {
-        de: "Wählen Sie die Art der Abholungen, die als Ereignis im Kalender angezeigt werden sollen",
-        en: "Select the type of pickups that you want to appear as events in the calendar"
-      }
-    ],
-    ["eventTimesTitle", {de: "Ereignis-Dauer", en: "Event Duration"}],
-    [
-      "eventTimesSubtitle", {
-        de: "Stellen Sie die Start- und Endzeit des Kalender-Eintrags ein",
-        en: "Set the start and end time of the calendar events"
-      }
-    ],
-    ["labelAllDay", {de: "Ganztätig", en: "Allday"}],
-    ["eventAlarmTitle", {de: "Ereignis-Erinnerung", en: "Event Alarm"}],
-    ["eventAlarmSubtitle", {de: "Stellen Sie eine Erinnerung vor dem Ereignis ein", en: "Set an alarm notification before the event"}],
-    ["labelAlarmMinutesBefore", {de: "Minute(n) vorher", en: "minute(s) before"}],
-    ["labelAlarmDisable", {de: "Erinnerung deaktivieren", en: "Disable Alarm"}],
-    ["eventOffsetTitle", {de: "Ereignis-Offset", en: "Offset Event"}],
-    ["eventOffsetSubtitle", {de: "Stellen Sie ein, ob Sie das Ereignis vor dem Abholungstag anzeigen möchten", en: "Set if you want to show the event before the actual pickup day"}],
-    ["labelOffsetDaysEarlier", {de: "Tag(e) vorher", en: "day(s) earlier"}],
-    [
-      "eventOffsetInfo", {
-      de: "Verwenden Sie dies, wenn Sie das Ereignis in Ihrem Kalender bereits vor der tatsächlichen Abholung angezeigt haben wollen. Dies kann nützlich sein, um sich bspw. am Abend vorher eine Erinnerung zu setzen, den Müll rechtzeitig bereitzustellen.", 
-      en: "Use this, if you want the event to appear in your calendar before the actual pickup happens. This can be useful e.g. to set yourself a reminder to put out the trash, e.g. the night before the pickup happens."}
-    ],
-    ["eventTitleTitle", {de: "Ereignis-Titel", en: "Event Title"}],
-    ["eventTitleSubtitle", {de: "Stellen Sie den Titel ein, wie das Ereignis im Kalender angezeigt werden soll", en: "Set the title that will be shown in the calendar"}],
-    ["labelEventTitle", {de: "Ereignis-Titel-Vorlage", en: "Event Title Template"}],
-    ["labelEventTitlePreview", {de: "Ereignis-Titel-Vorschau", en: "Event Title Preview"}],
-    ["eventTitleDefaultTitle", {de: "Abholung %1 (%2)", en: "Pickup %1 (%2)"}],
-    [
-      "eventTitlePlaceholderInfo", {
-      de: "Verwenden Sie '%1' und '%2' als Platzhalter für die Abholungsart und die Abholfrequenz", 
-      en: "You can use '%1' and '%2' as placeholders for the event type and the event schedule frequency."}
-    ],
-    ["yourIcalLink", {de: "Ihr benutzerdefinierter iCal-Link", en: "Your Custom iCal Link"}],
   ]);
-
-  const formatOptions = ref<UiICalFormatOptions>({
-    type: [],
-    startTime: "06:30",
-    endTime: "07:00",
-    alarm: 15,
-    allDay: false,
-    offsetEvent: 0,
-    customSummary: i18n.t("iCalGen.eventTitleDefaultTitle")
-  });
-
-  const disabledAlarm = ref<boolean>(false);
-
-  const apiUrl = (() => {
-    const url = useRequestURL();
-    url.pathname = "/api/v1/all";
-    return url;
-  })();
-
-  const icalUrl = computed( () => {
-    const urlParams = new URLSearchParams([
-      ["streetname", cookieStreet.value.streetname],
-      ["streetno", cookieStreet.value.streetno],
-      ["format", "ical"],
-      ["translate", cookieLanguage.value || "de"],
-    ]);
-
-    const optionalParams: [queryParam: string, queryValue: string | boolean | number | undefined][] = [
-      ["startTime", formatOptions.value.startTime],
-      ["endTime", formatOptions.value.endTime],
-      ["allDay", formatOptions.value.allDay],
-      ["alarm", (disabledAlarm.value) ? undefined : formatOptions.value.alarm || 0 * 60],
-      ["offsetEvent", formatOptions.value.offsetEvent],
-      ["customSummary", formatOptions.value.customSummary],
-      ["type", formatOptions.value.type?.join(",")]
-    ];
-
-    optionalParams.forEach(queryParamTuple => {
-      if (queryParamTuple[1]) {
-        urlParams.append(queryParamTuple[0], queryParamTuple[1].toString())
-      }
-    });
-
-    const localApiUrl = apiUrl;
-
-    localApiUrl.search = urlParams.toString();
-    return localApiUrl.toString();
-  })
-
-
-  const eventSummaryPreview = computed( () => {
-    return (formatOptions.value.customSummary && formatOptions.value.customSummary?.length > 0) 
-      ? formatOptions.value.customSummary?.replaceAll("%1", i18n.t("waste_residual")).replaceAll("%2", i18n.t("schedule_W2"))
-      : `${i18n.t("waste_residual")} (${i18n.t("schedule_W2")})`
-  });
-
 </script>
